feat(file-storage): rename util file, throw on missing env variables, log errors
